feat: default to Precision.fp32 and pin flutter_litert ^3.8.0; release 3.7.0
Adopts flutter_litert 3.8.0 and follows its precision default change.

This package has the strongest reason of any to make the switch. Pose models
were the worst fp16 offenders in the cross-platform matrix: pose_landmark_lite
missed CPU-reference parity by 159x the tolerance on Adreno 740, and
pose_landmark_full by 75 to 85x on every other GPU measured. Landmark outputs
are pixel-space coordinates and fp16 carries roughly three decimal digits of
mantissa, so that error lands directly on keypoint positions.

fp16 was the default at 11 call sites. The cost is real rather than free: fp32
is a median 29.9% slower on GPU across the five GPUs measured, with Apple M4
the lone exception at 6.5% faster. Every changed site was a default parameter
value, never an explicit call-site choice, so the parameter is retained and
callers can pass Precision.fp16 to restore the old behaviour.

Released as 3.7.0 rather than 3.6.1. The pending 3.6.1 was an example-only
engine-badge rename, but changing a precision default changes numeric output,
which is a minor-level change rather than a patch.

+++ lib/src/detector/pose_detector_native.dart
@@ -116,7 +116,7 @@ class PoseDetector {
     PerformanceConfig performanceConfig = const PerformanceConfig(),
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     bool enableSegmentation = false,
     // Web-only; accepted here for API parity but ignored on native.
     bool useLiteRt = true,
@@ -162,7 +162,7 @@ class PoseDetector {
     PerformanceConfig performanceConfig = const PerformanceConfig(),
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     bool enableSegmentation = false,
     // Web-only; accepted here for API parity but ignored on native.
     bool useLiteRt = true,
@@ -225,7 +225,7 @@ class PoseDetector {
     PerformanceConfig performanceConfig = const PerformanceConfig(),
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     bool enableSegmentation = false,
   }) async {
     if (isReady) {
